Trying the game engine

+++ Hexenspiel/src/main/java/trinat/hexenspiel/hexenspiel/Pumpkin.java
@@ -7,7 +7,7 @@ public class Pumpkin extends GameFigures {
     private Rectangle rectangle = super.getRectangle();
     private int speed = 8;
 
-    public Pumpkin(int sceneWidth, int scenePositionY, int sceneHeight) {
+    public Pumpkin(double sceneWidth, double scenePositionY, double sceneHeight) {
         super();
         this.rectangle.setHeight(20);
         this.rectangle.setWidth(20);
@@ -20,7 +20,7 @@ public class Pumpkin extends GameFigures {
 
     }
 
-    public Pumpkin(int sceneWidth, int scenePositionY, int sceneHeight,int positionX, int positionY, int Width, int Height, int ArcWidth, int ArcHeight) {
+    public Pumpkin(double sceneWidth, double scenePositionY, double sceneHeight,double positionX, double positionY, double Width, double Height, double ArcWidth, double ArcHeight) {
         super();
         this.rectangle.setX(positionX);
         this.rectangle.setY(RandomNumberGenerator.random(scenePositionY, scenePositionY + sceneHeight - this.rectangle.getHeight()));
